Initial refactor commit

+++ lib/Controller/ShiftsCalendarChangeController.php
@@ -1,10 +1,19 @@
 <?php
+/*
+ * @copyright Copyright (c) 2021. Fabian Kirchesch <fabian.kirchesch@csoc.de>
+ * @copyright Copyright (c) 2023. Kevin Küchler <kevin.kuechler@csoc.de>
+ *
+ * @author Fabian Kirchesch <fabian.kirchesch@csoc.de>
+ * @author Kevin Küchler <kevin.kuechler@csoc.de>
+ */
 
 
 namespace OCA\Shifts\Controller;
 
-
+use OCP\AppFramework\Http;
 use OCA\Shifts\AppInfo\Application;
+use OCA\Shifts\Settings\Settings;
+use OCA\Shifts\Service\PermissionService;
 use OCA\Shifts\Service\ShiftsCalendarChangeService;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\DataResponse;
@@ -15,16 +24,20 @@ class ShiftsCalendarChangeController extends Controller
 	/** @var ShiftsCalendarChangeService */
 	private $service;
 
-	/** @var string */
-	private $userId;
+	/** @var Settings */
+	private $settings;
+
+	/** @var PermissionService */
+	private $permService;
 
 	use Errors;
 
 
-	public function __construct(IRequest $request, ShiftsCalendarChangeService $service, $userId){
+	public function __construct(IRequest $request, ShiftsCalendarChangeService $service, Settings $settings, PermissionService $permService){
 		parent::__construct(Application::APP_ID, $request);
 		$this->service = $service;
-		$this->userId = $userId;
+		$this->settings = $settings;
+		$this->permService = $permService;
 	}
 
 	/**
@@ -32,6 +45,10 @@ class ShiftsCalendarChangeController extends Controller
 	 * @NoCSRFRequired
 	 */
 	public function index(): DataResponse {
+		if(!$this->permService->isRequestingUserAdmin()) {
+			return new DataResponse(NULL, Http::STATUS_UNAUTHORIZED);
+		}
+
 		return new DataResponse($this->service->findAll());
 	}
 
@@ -42,6 +59,10 @@ class ShiftsCalendarChangeController extends Controller
 	 * @return DataResponse
 	 */
 	public function show(int $id): DataResponse {
+		if(!$this->permService->isRequestingUserAdmin()) {
+			return new DataResponse(NULL, Http::STATUS_UNAUTHORIZED);
+		}
+
 		return $this->handleNotFound(function () use($id){
 			return $this->service->find($id);
 		});
@@ -61,8 +82,11 @@ class ShiftsCalendarChangeController extends Controller
 	 * @param bool $isDone
 	 * @return DataResponse
 	 */
-	public function create(int $shiftId, int $shiftTypeId, string $shiftDate, string $oldUserId, string $newUserId,
-						   string $action, string $dateChanged, string $adminId, bool $isDone): DataResponse {
+	public function create(int $shiftId, int $shiftTypeId, string $shiftDate, string $oldUserId, string $newUserId, string $action, string $dateChanged, string $adminId, bool $isDone): DataResponse {
+		if(!$this->permService->isRequestingUserAdmin()) {
+			return new DataResponse(NULL, Http::STATUS_UNAUTHORIZED);
+		}
+
 		return new DataResponse($this->service->create($shiftId, $shiftTypeId, $shiftDate, $oldUserId, $newUserId, $action, $dateChanged, $adminId, $isDone));
 	}
 
@@ -81,9 +105,12 @@ class ShiftsCalendarChangeController extends Controller
 	 * @param bool $isDone
 	 * @return DataResponse
 	 */
-	public function update(int $id,int $shiftId, int $shiftTypeId, string $shiftDate, string $oldUserId, string $newUserId,
-						   string $action, string $dateChanged, string $adminId, bool $isDone):DataResponse
+	public function update(int $id,int $shiftId, int $shiftTypeId, string $shiftDate, string $oldUserId, string $newUserId, string $action, string $dateChanged, string $adminId, bool $isDone):DataResponse
 	{
+		if(!$this->permService->isRequestingUserAdmin()) {
+			return new DataResponse(NULL, Http::STATUS_UNAUTHORIZED);
+		}
+
 		return $this->handleNotFound(function() use ($id, $shiftId, $shiftTypeId, $shiftDate, $oldUserId, $newUserId, $action, $dateChanged, $adminId, $isDone){
 			return $this->service->update($id, $shiftId, $shiftTypeId, $shiftDate, $oldUserId, $newUserId, $action, $dateChanged, $adminId, $isDone);
 		});
@@ -97,6 +124,10 @@ class ShiftsCalendarChangeController extends Controller
 	 */
 	public function destroy(int $id): DataResponse
 	{
+		if(!$this->permService->isRequestingUserAdmin()) {
+			return new DataResponse(NULL, Http::STATUS_UNAUTHORIZED);
+		}
+
 		return $this->handleNotFound(function() use($id) {
 			return $this->service->delete($id);
 		});
